+++ GameObject.h
@@ -16,7 +16,7 @@ typedef struct GameObject
 {
     int id;                 //id assigned on creation
     Vec3D pos, vel, acc;    //vectors initialised to ZERO
-    int rPos, rVel, rAcc;   //rotational properties initialised to 0
+    double rPos, rVel, rAcc;   //rotational properties initialised to 0
     Circle BCirc;             //bounding circle for collision detection
     int mass;
     bool isStationary;      //flag to indicate no need for physics update
@@ -32,9 +32,9 @@ int GO_getNumGO();
 void GO_setAcc(GameObject* go, Vec3D a);
 void GO_setVel(GameObject* go, Vec3D v);
 void GO_setPos(GameObject* go, Vec3D p);
-void GO_setRPos(GameObject* go, int p);
-void GO_setRVel(GameObject* go, int v);
-void GO_setRAcc(GameObject* go, int a);
+void GO_setRPos(GameObject* go, double p);
+void GO_setRVel(GameObject* go, double v);
+void GO_setRAcc(GameObject* go, double a);
 void GO_setStationary(GameObject* go, bool b);
 void GO_setBCirc(GameObject* go, Circle c);
 void GO_setMass(GameObject* go, int m);
@@ -43,9 +43,9 @@ void GO_setMass(GameObject* go, int m);
 Vec3D GO_getAcc(const GameObject* go);
 Vec3D GO_getVel(const GameObject* go);
 Vec3D GO_getPos(const GameObject* go);
-int GO_getRPos(const GameObject* go);
-int GO_getRVel(const GameObject* go);
-int GO_getRAcc(const GameObject* go);
+double GO_getRPos(const GameObject* go);
+double GO_getRVel(const GameObject* go);
+double GO_getRAcc(const GameObject* go);
 int GO_getMass(const GameObject* go);
 bool GO_isStationary(const GameObject* go);
